Refactoring:  - Moved all extension functions and properties targeting `TreeView`    from `Nodes.kt` to `TreeView.kt`
